Added 2D and 3D functionality for US map; this enables us to switch between 2d and 3d
